refactor: abstract `dayjs` into its own package
