+++ apps/web/package.json
@@ -14,6 +14,7 @@
   },
   "dependencies": {
     "@blobscan/api": "0.0.1",
+    "@blobscan/dayjs": "^0.0.1",
     "@blobscan/db": "0.0.1",
     "@blobscan/tailwind-config": "0.0.1",
     "@fontsource/inter": "^4.5.15",
@@ -29,7 +30,6 @@
     "@trpc/react-query": "^10.23.0",
     "@trpc/server": "^10.23.0",
     "classnames": "^2.3.2",
-    "dayjs": "^1.11.7",
     "echarts": "^5.4.2",
     "echarts-for-react": "^3.0.2",
     "ethers": "^5.7.2",
