+++ apps/web/src/components/Cards/SurfaceCards/BlobTransactionCard.tsx
@@ -15,7 +15,8 @@ import { animated, useSpring } from "@react-spring/web";
 import "react-loading-skeleton/dist/skeleton.css";
 import Skeleton from "react-loading-skeleton";
 
-import dayjs from "~/dayjs";
+import dayjs from "@blobscan/dayjs";
+
 import { type Block } from "~/types";
 import {
   buildAddressRoute,
