+++ packages/api/src/clients/google.ts
@@ -15,7 +15,7 @@ const credentials =
     : undefined;
 
 export function getStorage() {
-  if (env.NODE_ENV === "development") {
+  if (process.env.NODE_ENV === "development") {
     return new Storage({ apiEndpoint: "http://localhost:4443" });
   } else {
     return new Storage(credentials ? { credentials } : undefined);
