+++ packages/api/src/middlewares/withDates.ts
@@ -1,6 +1,6 @@
-import dayjs from "dayjs";
 import { z } from "zod";
 
+import dayjs from "@blobscan/dayjs";
 import { normalizeDate, type DatePeriod } from "@blobscan/db";
 
 import { t } from "../clients/trpc";
