+++ packages/db/package.json
@@ -19,11 +19,11 @@
     "seed": "ts-node prisma/seed.ts"
   },
   "dependencies": {
+    "@blobscan/dayjs": "^0.0.1",
     "@prisma/client": "^4.12.0"
   },
   "devDependencies": {
     "@faker-js/faker": "^8.0.2",
-    "dayjs": "^1.11.7",
     "js-sha256": "^0.9.0",
     "prisma": "^4.12.0",
     "typescript": "^5.0.3"
